backend/requirements.txt
torch
torchvision
fastai
huggingface_hub
pillow
scikit-image
matplotlib
